Increase git depth for TileDB conda recipe
If a commit is merged while the nightly CI jobs are running, this can
break the nightly build. It just happened! Only one of the linux-64
jobs was able to start before the commit was merged, and
then all the other jobs failed since that commit no longer existed
in the shallow clone

https://dev.azure.com/jdblischak/feedstock-builds/_build/results?buildId=68&view=results

Sun Oct 9 13:59:30 2022 -0400:
https://github.com/TileDB-Inc/TileDB/commit/5f0f5c41facd86b7e44203e0375d3157d8dc6aa3

Sun Oct 9 21:39:26 2022 -0400
https://github.com/TileDB-Inc/TileDB/commit/2f3b15c088fd8e2c7e81c38dfaa9540b09b2a87f